+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to `angular-rave` will be documented here
 
+## 3.0.0 - 2020-09-13
+### Changes
+- `forRoot` configuration no longer accepts an object. Now accepts only the public key
+- `PUBKey` now `public_key`
+- Implemented updated to `flutterwave's` inline script.
+- `Customization` and `customer` object 
+
+### Removed
+- Removed the `autoClose` option
+
+### Fixed
+- No reload on successful transaction
+- Callback is now being called after a successful transaction
+
 ## 2.0.0 - 2019-10-19
 ### Added
 - Module `forRoot` configuration